Support ListDDLPublishRecords API.

+++ aliyun-java-sdk-dms-enterprise/src/main/java/com/aliyuncs/dms_enterprise/model/v20181101/GetMetaTableDetailInfoResponse.java
@@ -111,7 +111,7 @@ public class GetMetaTableDetailInfoResponse extends AcsResponse {
 
 			private Boolean autoIncrement;
 
-			private Integer dataLength;
+			private Long dataLength;
 
 			private Integer dataScale;
 
@@ -167,11 +167,11 @@ public class GetMetaTableDetailInfoResponse extends AcsResponse {
 				this.autoIncrement = autoIncrement;
 			}
 
-			public Integer getDataLength() {
+			public Long getDataLength() {
 				return this.dataLength;
 			}
 
-			public void setDataLength(Integer dataLength) {
+			public void setDataLength(Long dataLength) {
 				this.dataLength = dataLength;
 			}
